+++ json_serializable/test/json_serializable_test.dart
@@ -219,4 +219,7 @@ const _expectedSchemaTests = {
   'JsonSchemaNullableTest',
   'JsonSchemaNestedTest',
   'JsonSchemaNonCollectionTest',
+  'JsonSchemaGetterTest',
+  'JsonSchemaRecursiveListTest',
+  'JsonSchemaRecursiveListIssue',
 };