+++ json_serializable/lib/src/utils.dart
@@ -71,7 +71,6 @@ JsonSerializable _valueForAnnotation(ConstantReader reader) => JsonSerializable(
       reader.read('disallowUnrecognizedKeys').literalValue as bool?,
   explicitToJson: reader.read('explicitToJson').literalValue as bool?,
   fieldRename: readEnum(reader.read('fieldRename'), FieldRename.values),
-  enumFieldRename: readEnum(reader.read('enumFieldRename'), FieldRename.values),
   genericArgumentFactories:
       reader.read('genericArgumentFactories').literalValue as bool?,
   ignoreUnannotated: reader.read('ignoreUnannotated').literalValue as bool?,