+++ json_serializable/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 6.13.0-wip
 
+- Add `JsonSerializable.enumFieldRename` configuration option.
+  ([#1556](https://github.com/google/json_serializable.dart/pull/1556))
 - Support `JsonSerializable.dateTimeUtc` configuration option. 
   ([#1371](https://github.com/google/json_serializable.dart/issues/1371))
 - Remove `json_schema` from `dependencies` (it is only used in tests).